move function to util file
